django_docker/requirements.txt
django
gunicorn
djangorestframework
django-rest-auth